[MOOSE-12]: move admin config and common js partials out of admin specific structure
