+++ wp-content/themes/core/assets/js/common/resize.js
@@ -4,7 +4,7 @@
  * @description Kicks in any third party plugins that operate on a sitewide basis.
  */
 
-import { triggerCustomEvent } from 'utils/events';
+import { triggerCustomEvent } from 'utils/events.js';
 import viewportDims from './viewport-dims';
 
 const resize = () => {
