+++ wp-content/themes/core/assets/js/config/options.js
