+++ wp-content/themes/core/assets/js/config/state.js
